LayoutTests/http/tests/media/media-stream/enumerate-devices-iframe-allow-attribute-expected.txt

CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call enumerateDevices.
CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call enumerateDevices.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call enumerateDevices.
CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute 'microphone'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Permission policy 'Microphone' check failed for element with origin 'http://localhost:8000' and allow attribute 'camera'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute 'microphone;speaker-selection'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.


PASS Wait for frames to be loaded
PASS none
PASS microphone
PASS camera
PASS camera+microphone
PASS microphone+speakerselection
PASS camera+microphone+speakerselection
PASS same-origin
